Add unoptimised simple .wasm file, containing plenty of straight forward calls
Should be a useful example wasm for watching and learning how the wasm
call stack works in the WASim VM

The .wasm file was generated in three steps:

1. Generate an initial .wasm file using TinyGo

    $ tinygo build -target wasm -no-debug -opt=1 -gc dumb -o simple-gc-dumb.wasm simple.go

Note that this uses a special version of TinyGo, modified to not
include unneeded function calls, which reduces the resulting file
size:

    https://github.com/justinclift/tinygo/commit/8a5665cb1ba5e454790b03d817547e7997ec2e7e

2. Convert the .wasm file to .wat

    $ wasm2wat --generate-names -o simple-gc-dumb.wat simple-gc-dumb.wasm

3. Convert the .wat file back to .wasm

    $ wat2wasm -o simple-gc-dumb.wasm simple-gc-dumb.wat

This recreation of the .wasm file from the .wat reduces the file size
a bit more, without running an optimisation pass, as it removes the
names of the functions from the end of the file.

Those function names are added by TinyGo (maybe due to LLVM), as a
custom section on the end of the file, and aren't useful for the
current objective.